+++ src/components/pages/teams/TeamsContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import Teams from './Teams';
-import { getTeams, createTeam, updateTeam, setEditing } from '../../../modules/teams';
+import { getTeams } from '../../../modules/teams';
 import { includes } from '../../../helpers/utils';
 
 const isManager = (team, userId) => {
@@ -15,7 +15,7 @@ const belongsToTeam = (team, userId) => {
   return (isManager(team, userId) || isUser(team, userId));
 }
 
-const mapStateToProps = ({ areas, user, teams }) => {
+const mapStateToProps = ({ user, teams }) => {
   const userId = user.data.id;
   let teamId = null;
   if(teams.ids.length > 0){
@@ -23,15 +23,10 @@ const mapStateToProps = ({ areas, user, teams }) => {
         return belongsToTeam(teams.data[key], userId);
       }) || null;
   }
-  const areaValues = Object.keys(areas.areas).map((key) => ({
-    value: areas.areas[key].id,
-    label: areas.areas[key].attributes.name
-  }));
+
   const team = teams.data[teamId]
     return { 
-      userId: user.data.id,
       team,
-      areaValues,
       isManager: team && isManager(team, userId),
       editing: teams.editing
     };
@@ -41,15 +36,6 @@ const mapStateToProps = ({ areas, user, teams }) => {
    return {
      getTeams: () => {
        dispatch(getTeams());
-    },
-    createTeam: (team) => {
-       dispatch(createTeam(team));
-    },
-    updateTeam: (team, id) => {
-       dispatch(updateTeam(team, id));
-    },
-    setEditing: (value) => {
-       dispatch(setEditing(value));
     }
   }
  }
